refactor(ui): address Gemini code review — extract TopDownRenderer, fix perf
Address all 6 Gemini code review comments on PR #127:

HIGH:
- Hoist depthScale minY/maxY/rangeY computation outside fixture loop
  in AudienceView (was O(N²), now O(N))
- Extract shared TopDownRenderer object from duplicated drawing code
  in VenueCanvas and TopDownEditor (~300 lines deduplicated)

MEDIUM:
- Make zLevels collection idiomatic Kotlin (map + toSet)
- Re-add brightness variation comments in bar segment rendering
